requirements.txt
Flask
google-generativeai
python-docx
PyMuPDF
python-dotenv
gunicorn
